+++ lib/Cardano/Address/Crypto/ListN.hs
@@ -6,6 +6,8 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 
+{-# OPTIONS_GHC -Wno-dodgy-imports #-}
+
 -- | Vendored minimal sized-list from @basement@ (BSD-3-Clause).
 -- Only the subset used by BIP39 encoding.
 module Cardano.Address.Crypto.ListN
@@ -15,7 +17,7 @@ module Cardano.Address.Crypto.ListN
     , toListN_
     , map
     , mapM
-    , foldl'
+    , Cardano.Address.Crypto.ListN.foldl'
     ) where
 
 import Prelude hiding