+++ lib/Cardano/Address/Script.hs
@@ -105,8 +105,6 @@ import Data.Text
     ( Text )
 import Data.Traversable
     ( for )
-import Data.Word
-    ( Word8 )
 import GHC.Generics
     ( Generic )
 import Numeric.Natural
@@ -134,7 +132,7 @@ data Script (elem :: Type)
     = RequireSignatureOf !elem
     | RequireAllOf ![Script elem]
     | RequireAnyOf ![Script elem]
-    | RequireSomeOf Word8 ![Script elem]
+    | RequireSomeOf Word ![Script elem]
     | ActiveFromSlot Natural
     | ActiveUntilSlot Natural
     deriving stock (Generic, Show, Eq)
@@ -190,7 +188,7 @@ serializeScript script =
 -- | Represents the cosigner of the script, ie., party that co-shares the script.
 --
 -- @since 3.2.0
-newtype Cosigner = Cosigner Word8
+newtype Cosigner = Cosigner Word
     deriving (Generic, Ord, Eq)
 instance Hashable Cosigner
 instance NFData Cosigner
@@ -474,6 +472,7 @@ recommendedValidation = \case
 
     RequireSomeOf m script -> do
         when (m == 0) $ Left MZero
+        when (m > 255) $ Left BigNInRequiredSomeOf
         when (length (omitTimelocks script) < fromIntegral m) $ Left ListTooSmall
         when (hasDuplicate script) $ Left DuplicateSignatures
         when (redundantTimelocks script) $ Left RedundantTimelocks
@@ -537,6 +536,8 @@ validateScriptTemplate level (ScriptTemplate cosigners_ script) = do
     check DuplicateXPubs (Set.size cosignerKeys == Map.size cosigners_)
     check UnknownCosigner (cosignerSet `Set.isSubsetOf` scriptCosigners)
     check MissingCosignerXPub (scriptCosigners `Set.isSubsetOf` cosignerSet)
+    when (level == RecommendedValidation) $
+        check BigCosignerNumber (all (\(Cosigner n) -> n <= 255) scriptCosigners)
   where
     scriptCosigners = Set.fromList $ foldScript (:) [] script
     cosignerKeys = Set.fromList $ Map.elems cosigners_
@@ -579,6 +580,7 @@ data ErrRecommendedValidateScript
     | DuplicateSignatures
     | RedundantTimelocks
     | TimelockTrap
+    | BigNInRequiredSomeOf
     deriving (Eq, Show)
 
 -- | Possible validation errors when validating a script template
@@ -591,6 +593,7 @@ data ErrValidateScriptTemplate
     | MissingCosignerXPub
     | NoCosignerInScript
     | NoCosignerXPub
+    | BigCosignerNumber
     deriving (Eq, Show)
 
 -- | Pretty-print a script validation error.
@@ -628,6 +631,8 @@ prettyErrValidateScript = \case
         "Some timelocks used are redundant (which is not recommended)."
     NotRecommended TimelockTrap ->
         "The timelocks used are contradictory when used with 'all' (which is not recommended)."
+    NotRecommended BigNInRequiredSomeOf ->
+        "The script requires more than 255 signatures. Such scripts might result in transactions exceeding the possible transaction size limit (which is not recommended)."
 
 -- | Pretty-print a script template validation error.
 --
@@ -647,6 +652,8 @@ prettyErrValidateScriptTemplate = \case
         "The script template must have at least one cosigner with an extended public key."
     UnknownCosigner ->
         "The specified cosigner must be present in the script of the template."
+    BigCosignerNumber ->
+        "The script template contains a cosigner number greater than 255. Such scripts might result in transactions exceeding the possible transaction size limit (which is not recommended)."
 --
 -- Internal
 --
@@ -791,11 +798,12 @@ instance ToJSON Cosigner where
 
 instance FromJSON Cosigner where
     parseJSON = withText "Cosigner" $ \txt -> case T.splitOn "cosigner#" txt of
-        ["",numTxt] ->  case T.decimal numTxt of
+        ["",numTxt] ->  case T.decimal @Integer numTxt of
             Right (num,"") -> do
-                when (num < minBound @Word8 || num > maxBound @Word8) $
-                        fail "Cosigner number should be between '0' and '255'"
-                pure $ Cosigner num
+                let maxWord = toInteger (maxBound @Word)
+                when (num > maxWord) $
+                        fail "Cosigner number is out of bounds"
+                pure $ Cosigner (fromInteger num)
             _ -> fail "Cosigner should be enumerated with number"
         _ -> fail "Cosigner should be of the form: cosigner#num"