+++ lib/Cardano/Address/Script/Parser.hs
@@ -34,8 +34,6 @@ import Data.Char
     ( isDigit, isLetter )
 import Data.Text
     ( Text )
-import Data.Word
-    ( Word8 )
 import Numeric.Natural
     ( Natural )
 import Text.ParserCombinators.ReadP
@@ -164,7 +162,7 @@ activeUntilSlotParser = do
     _identifier <- P.string "active_until"
     ActiveUntilSlot <$> slotParser
 
-naturalParser :: ReadP Word8
+naturalParser :: ReadP Word
 naturalParser = do
     P.skipSpaces
     fromInteger . read <$> P.munch1 isDigit